pyproject.toml
[project]
name = "tmGrammars"
version = "0.0.1"
description = "A python tool to manipulate and run TextMate Grammars"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
tmgc = "tmGrammars.cli:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.2.1",
    "pytest-asyncio>=0.20.3",
    "aiofiles>=23.1.0",
    "pytest-mock>=3.10.0",
]
